+++ microservices/dep_check/scan.py
